% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{calcIVT-package}
\alias{calcIVT}
\alias{calcIVT-package}
\title{Functions for calculating Intermediate Value Tables (IVTs)}
\description{
Functions for calculating Intermediate Value Tables (IVTs).  This is a
high-speed implementation of the Morning Report Algorithms developed by
Brett Amidan and Tom Ferryman.
}
\details{
\tabular{ll}{
Package: \tab calcIVT\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2014-10-02\cr
License: \tab LICENSE\cr }
}
\author{
Landon Sego, Brett Amidan, Tom Ferryman

Maintainer: Landon Sego <Landon.Sego@pnnl.gov>
}
\keyword{package}

