% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calcIVTslow}
\alias{calcIVTslow}
\title{Fits the moving window quadratic regression model and summarizes results by
phases}
\usage{
calcIVTslow(y, phase.vec, x1, min.window = 5, start = 1, skip = 1,
  actual.min.max = TRUE, linear.only = FALSE)
}
\arguments{
\item{y}{Vector of data for a particular variable for the whole flight}

\item{phase.vec}{Phase vector that indicates the phases of the flight (must
have same length as \code{y} with elements that are in sequentially
increasing order)}

\item{x1}{See \code{\link{fitQ}}}

\item{min.window}{See \code{\link{fitQ}}}

\item{start}{See \code{\link{fitQ}}}

\item{skip}{See \code{\link{fitQ}}}

\item{actual.min.max}{if \code{TRUE}, the actual minimum and maximum data
values within the phase are reported for \code{a.min} and \code{a.max}.  If
\code{FALSE}, The minimum and maximum values of the calculated intercepts
within the phase are reported for \code{a.min} and \code{a.max}.}

\item{linear.only}{See \code{\link{fitQ}}}
}
\value{
A named vector containing the coefficients of the regression fits
summarized by phase.  The names of the vector have the following general
format: \code{p.PHASE.COEFFICIENT.STATISTIC}, as illustrated in the
examples below: \item{p.1.a.min}{The minimum intercept (or minimum data
value if \code{actual.min.max=TRUE}) in phase 1} \item{p.4.a.max}{The
maximum intercept (or maximum data value if \code{actual.min.max=TRUE}) in
phase 4} \item{p.2.b.total}{The sum of the coefficients of the linear term
for phase 2} \item{p.5.c.ss}{The sum of squares of the coefficients of the
quadratic terms for phase 5} \item{p.1.d.total}{The sum of the RMSE's of
the regression fits in phase 1} \item{p.7.start.value}{The first data point
in phase 7} \item{p.8.end.value}{The last data point in phase 8}
\item{p.3.n.count}{The number of data points in phase 3}
}
\description{
A very slow version of \code{\link{calcIVTc}} that was coded independently
in order to check \code{\link{calcIVTc}}.  It is identical in every respect
to \code{\link{calcIVTc}} except that it uses a slower algorithm
that uses \code{\link{lm}} to fit the linear model across the window and it
returns the labeled IVT vector in a different order.
}
\details{
A test that verified that calcIVTslow() and calcIVTc() were equivalent was performred January 30, 2008
}
\author{
Landon Sego
}
\keyword{misc}

