% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/check_fitQ_args.R
\name{check_fitQ_args}
\alias{check_fitQ_args}
\title{Check and validate arguments for fitQ()}
\usage{
check_fitQ_args(x1 = -10:10, min.window = 5, start = 1, skip = 1,
  linear.only = FALSE)
}
\arguments{
\item{x1}{The predictor of the regression model, whose length must
be odd, and it must be monotonic (increasing or decreasing).  Typically this
would be an evenly spaced, increasing vector.}

\item{min.window}{The minimum number of non-missing data points in a window
that are required to fit the regression model.}

\item{start}{The index of the center of the first window}

\item{skip}{The number of indexes to advance the center of the moving
window each time the model is fit.}

\item{linear.only}{\code{=TRUE} fits a simple linear regression model with
\code{x1} as the single predictor, instead of a quadratic regression model.}
}
\value{
An object of class \code{valid_fitQ_args} that can be passed to the \code{x1} argument
of \code{\link{fitQ}}. It is a list with the same
elements that were passed into the function: \code{x1}, \code{min.window}, \code{start},
\code{skip}, \code{linear.only}, plus one more, the bandwidth (\code{bw}), which is equal to
\code{(length(x1) - 1) / 2}.
}
\description{
Check and validate arguments for \code{\link{fitQ}} before they are used.
}
\details{
This function exists so that multiple calls to \code{\link{fitQ}}
from \code{\link{getFeatures}} would not involve repeating the checks on essentially
the same arguments.
}
\author{
Landon Sego
}

