% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/check_getFeatures_args.R
\name{check_getFeatures_args}
\alias{check_getFeatures_args}
\title{Check and validate arguments for getFeatures() and ddply_getFeatures()}
\usage{
check_getFeatures_args(y, cont = NULL, disc = NULL, centerScale = TRUE,
  stats = c("min", "q1", "mean", "med", "q3", "max", "sd", "count"),
  fitQargs = NULL)
}
\arguments{
\item{y}{Data frame, each row containing a vector of measurements for a particular point in time,
with columns indicating the measured variables (and possibly other descriptive variables).  The
data processed presuming the rows are orderd chronologically.}

\item{cont}{Vector of integers or a character vector indicating the columns
of \code{x} that correspond to continuous variables.  These are the variables from which features
will be extracted by fitting the moving regression model using \code{\link{fitQ}}.}

\item{disc}{Vector of integers or character vector indicating the columns
of \code{x} that correspond to variables that will be treated as discrete. These are the variables
from which features will be extracted using \code{\link{discFeatures}}.}

\item{centerScale}{Logical indicating whether the continuous variables (indicate by \code{cont})
should be centered and scaled by the global mean and standard deviation of that variable.  By
'global', we mean all the values of a continuous variable, say \code{x}, in \code{y} are used
to compute the mean and standard deviation.  The resulting value for the continuous variable,
\code{x}, is equivalent to \code{y$x <- (y$x - mean(y$x)) / sd(y$x)}.}

\item{stats}{This argument defines the summary statistics that will be calculated
for each of the regression parameters.  It can be a character vector of summary statistics,
which are passed to \code{\link{summaryStats}}.  Or the function object returned by
\code{\link{summaryStats}} may be supplied.}

\item{fitQargs}{Named list of arguments for \code{\link{fitQ}}. If \code{NULL}, the default
arguments of \code{\link{fitQ}} are used. Any argument for \code{\link{fitQ}} may be included
except \code{y}.}
}
\value{
An object of class \code{valid_getFeatures_args} that can be passed to the first argument
of \code{\link{getFeatures}} or \code{\link{ddply_getFeatures}}. It is a list with the same
elements that were passed into the function: \code{y}, \code{cont}, \code{disc}, \code{stats},
and \code{fitQ}, with the exception that the continuous variables of \code{y} are now
centered and scaled if \code{centerScale = TRUE}, and \code{cont} and \code{disc} are converted
to the character column names (if integers were originally provided).
}
\description{
Check and validate arguments for \code{\link{getFeatures}} and \code{\link{ddply_getFeatures}}
before they are used by those functions.
}
\details{
This function exists so that multiple calls to \code{\link{getFeatures}}
from \code{\link{ddply_getFeatures}} would not involve repeating the checks on essentially
the same arguments.
}
\author{
Landon Sego
}

