% Generated by roxygen2 (4.0.1): do not edit by hand
\name{continuousIVT}
\alias{continuousIVT}
\title{Wrapper function that calculates the IVTs for multiple continuous variables}
\usage{
continuousIVT(cNames, phase.vec, phases.to.process, seconds, nominal.skip = 1,
  nominal.bw = 5, actual.min.max = TRUE)
}
\arguments{
\item{cNames}{character vector containing the objects names (in the global environment)
of the continuous vectors, each vector corresponding to a variable (or flight
parameter)}

\item{phase.vec}{Phase vector (usually integer valued) that indicates the
phases of the flight, with one element for each second of the flight.  For
variables with Hz other than 1, the \code{phase.vec} is modified
accordingly.}

\item{phases.to.process}{numeric vector of contiguous phases to be
processed}

\item{seconds}{number of seconds in the flight}

\item{nominal.skip}{the nominal number of _seconds_ to slide the window for
each fitting}

\item{nominal.bw}{the nominal number of _seconds_ for the bandwidth, where
the bandwidth is approximately half the length of the moving window (length
of window = 2*bandwidth + 1)}

\item{actual.min.max}{see \code{\link{calcIVTc}}}
}
\value{
The continuous IVT matrix, with colnames corresponding to the
variables (parameters) requested in \code{cNames}, and rownames indicating
the phase, the coefficient of the quadratic model, and the summary
statistic, as described in \code{\link{calcIVTc}}.
}
\description{
Wrapper function for \code{\link{calcIVTc}} that calculates the
IVTs for multiple continuous variables
}
\author{
Landon Sego
}
\seealso{
\code{\link{calcIVTc}}
}
\keyword{misc}

