% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/demoData.R
\docType{data}
\name{demoData}
\alias{demoData}
\title{Some fake demonstration data to illustrate the qFeature package}
\format{'data.frame':	646 obs. of  11 variables.  See example below.}
\description{
Some fake demonstration data to illustrate the qFeature package
}
\examples{
data(demoData)
str(demoData)
head(demoData)
rm(demoData, envir = .GlobalEnv)
}
\keyword{datasets}

