% Generated by roxygen2 (4.0.2): do not edit by hand
\name{discFeatures}
\alias{discFeatures}
\title{Summarizes the transitions between discrete states by phase}
\usage{
discFeatures(y)
}
\arguments{
\item{y}{A vector with discrete states}
}
\value{
A named vector with the fraction of time spent in each state, and the
number of transitions from one state to another
}
\description{
Calculates the percentage of time in each state and the number transitions
between states
}
\examples{
x <- c("a","a","a","b","b","a","a","c","c","b","b")
discFeatures(x)

discFeatures(c(T, F, F, T, F, F, F))

discFeatures(c(1, 1, 1, 2, 2))

discFeatures(c(rep(pi, 3), rep(exp(1), 3), rep(sqrt(2), 3)))
}
\author{
Landon Sego
}
\seealso{
\code{\link{transMap}}, \code{\link{sp.table}}
}

