% Generated by roxygen2 (4.0.1): do not edit by hand
\name{discreteIVT}
\alias{discreteIVT}
\title{Wrapper function that calculates IVTs for multiple discrete variables}
\usage{
discreteIVT(dNames, phase.vec, phases.to.process, seconds)
}
\arguments{
\item{dNames}{character vector containing the objects names (in the global
environment) of the discrete
vectors, each vector corresponding to a discrete variable (or flight
parameter)}

\item{phase.vec}{Phase vector (usually integer valued) that indicates the
phases of the flight, with one element for each second of the flight.  For
variables with Hz other than 1, the \code{phase.vec} is modified
accordingly.}

\item{phases.to.process}{numeric vector of contiguous phases to be
processed}

\item{seconds}{number of seconds in the flight}
}
\value{
Discrete IVT matrix, with phases on the rows and parameters with
their various state transition/percentage summaries on the columns.
}
\description{
Wrapper function for \code{\link{discreteTransition}} that calculates
IVTs for multiple discrete variables
}
\details{
Discrete IVTs are created by tabulating the number of transitions made from
one discrete state to another (summarized by phase) and by calculating the
percentage of time spent within each state (again summarized by phase).
}
\author{
Landon Sego
}
\seealso{
\code{\link{discreteTransition}}
}
\keyword{misc}

