% Generated by roxygen2 (4.0.1): do not edit by hand
\name{discreteTransition}
\alias{discreteTransition}
\title{Summarizes the transitions between discrete states by phase}
\usage{
discreteTransition(y, phase.vec, phases.to.process, vname = "V")
}
\arguments{
\item{y}{an integer vector of a discrete flight parameter}

\item{phase.vec}{see \code{\link{continuousIVT}}}

\item{phases.to.process}{see \code{\link{continuousIVT}}}

\item{vname}{character string giving a prefix for the state transition
labels}
}
\value{
A matrix with the phases on the rows and the summaries of the
transition counts and the percentage of time spent in each state on the
columns.
}
\description{
Calculates the percentage of time in each state and the number transitions
between states, summarized by phase
}
\examples{
x <- c(1,1,1,2,2,1,1,3,3,2,2)
phase.vec <- c(rep(4,5), rep(5,6))
discreteTransition(x, phase.vec, 4:5, vname="Test")
}
\author{
Landon Sego
}
\seealso{
\code{\link{transMap}}, \code{\link{sp.table}}
}
\keyword{misc}

