% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fitQ}
\alias{fitQ}
\title{Fits a moving window quadratic (or simple linear) regression model}
\usage{
fitQ(y, x1, min.window = 5, start = 1, skip = 1, linear.only = FALSE)
}
\arguments{
\item{y}{A numeric vector}

\item{x1}{The linear predictor of a the regression model, whose length must
be odd, and it must be monotonic (increasing or decreasing).}

\item{min.window}{The minimum number of non-missing data points in a window
that are required to fit the regression model.}

\item{start}{The index of the center of the first window}

\item{skip}{The number of indexes to advance the center of the moving
window each time the model is fit.}

\item{linear.only}{\code{=TRUE} fits a simple linear regression model with
\code{x1} as the single predictor, instead of a quadratic regression model.}
}
\value{
A list with the 4 component vectors that contain the results of the
quadratic model fits.  \item{a}{The estimated intercepts} \item{b}{The
estimated linear coefficients} \item{c}{The estimated quadratic
coefficients.  These are \code{NA} is \code{linear.only=TRUE}} \item{d}{The
root mean squared error (RMSE)}
}
\description{
Fits a moving quadratic (or simple linear) regression model over a series
using a 2-sided window.  It dynamically accounts for the incomplete windows
which are caused by missing values and which occur at the beginning and end
of the series.
}
\details{
Fits the model

\code{y = a + b*x1 + c*x1^2 + e}

if \code{linear.only=FALSE}, or, if \code{linear.only=TRUE}, fits

\code{y = a + b*x1 + e}

The model is fit repeatedly over a moving window. For the quadratic model,
it rapidly calculates the least squares estimates by centering the
predictors so that they are orthogonal (or nearly orthogonal, if data in
the window are missing).  Then the orthogonal coefficients are
'backtransformed' to the original, non-orthogonal parameterization.
Centering and orthogonality are not required to quickly fit the simple
linear regression model.

The handling of the moving windows and missing values in this function is
very similar to \code{\link[Smisc]{smartFilter}} in \pkg{Smisc}.
}
\author{
Landon Sego
}
\seealso{
\code{\link{calcIVTc}}
}
\keyword{misc}

