% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fitQslow}
\alias{fitQslow}
\title{Fits the moving window quadratic (or simple linear) regression model}
\usage{
fitQslow(y, x1, min.window = 5, start = 1, skip = 1,
  linear.only = FALSE)
}
\arguments{
\item{y}{A numeric vector}

\item{x1}{The linear predictor of a the regression model, whose length must
be odd, and it must be monotonic (increasing or decreasing).}

\item{min.window}{The minimum number of non-missing data points in a window
that are required to fit the regression model.}

\item{start}{The index of the center of the first window}

\item{skip}{The number of indexes to advance the center of the moving
window each time the model is fit.}

\item{linear.only}{\code{=TRUE} fits a simple linear regression model with
\code{x1} as the single predictor, instead of a quadratic regression model.}
}
\value{
A list with the 4 component vectors that contain the results of the
quadratic model fits.  \item{a}{The estimated intercepts} \item{b}{The
estimated linear coefficients} \item{c}{The estimated quadratic
coefficients.  These are \code{NA} is \code{linear.only=TRUE}} \item{d}{The
root mean squared error (RMSE)}
}
\description{
A very slow version of \code{\link{fitQ}} that was coded independently in
order to check \code{\link{fitQ}}.  It is identical in every respect to
\code{\link{fitQ}} except that it uses a slower algorithm that uses \code{\link{lm}}
for the model fit.
}
\author{
Landon Sego
}
\keyword{misc}

