% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genDemoData.R
\name{genDemoData}
\alias{genDemoData}
\title{Generate the demo data for the qFeature package}
\usage{
genDemoData()
}
\value{
'data.frame':	646 obs. of  11 variables.  See example below.
}
\description{
Generate the demo data for the qFeature package
}
\details{
A call to \code{genDemoData} produces the same data as that given by \code{data(demoData)}. This approach allows us
to export this function and generate the data easily from other packages via \code{qFeature::genDemoData}
}
\examples{
# Generate the data
x <- genDemoData()
str(x)

# Compare to loading via data()
data(demoData)
identical(demoData, x)

# Remove the objects
rm(x, demoData)
}
\author{
Landon Sego
}

