% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getFeatures}
\alias{getFeatures}
\title{Fits the moving window regression model and summarizes results by phases}
\usage{
getFeatures(y, cont = NULL, disc = NULL, stats = c("min", "q1", "mean",
  "med", "q3", "max", "sd", "count"), ...)
}
\arguments{
\item{cont}{Vector of integers or character vector indicating the columns
of \code{x} that correspond to variables that will be treated as continuous}

\item{disc}{Vector of integers or character vector indicating the columns
of \code{x} that correspond to variables that will be treated as discrete}

\item{stats}{Character vector of summary statistics, which are passed to
\code{\link{summaryStats}}.}

\item{x}{Data frame}

\item{\dots}{Arguments for \code{\link{fitQ}}.}
}
\value{
A named vector containing the features for each of the variables
requested in \code{cont} and \code{disc}.  The names follow the form
[varname].[description], where the [varname] is specified in \code{cont} and
\code{disc}, and [description] follows the naming convention produced by
code{\link{summary.fitQ}} \code{\link{discFeatures}}.
}
\description{
Fits the moving window regression model over a single vector of data and
summarizes results by phases
}
\details{
A least one of \code{cont} or \code{disc} must be specified.
}
\author{
Landon Sego
}

