% Generated by roxygen2 (4.0.1): do not edit by hand
\name{identifyHz}
\alias{identifyHz}
\title{Identifies the frequency (Hz) of measurement of one or more variables}
\usage{
identifyHz(vNames, seconds, acceptable = c(1/32, 1/16, 1/8, 1/4, 1/2, 1, 2, 4,
  8, 16, 32))
}
\arguments{
\item{vNames}{a character vector containing the names of the data vectors
for which this function determines their Hz}

\item{seconds}{a numeric value indicating the number of seconds in the
flight}

\item{acceptable}{a vector indicating the acceptable Hz values.}
}
\value{
A named vector of the Hz for each variable.  The names correspond
to the variable names provided by \code{vnames}.
}
\description{
Identifies the frequency (Hz) of measurement of one or more variables
}
\details{
If the length of a each data vector is not a Hz multiple of the number of
seconds, then it is set to NA.
}
\author{
Landon Sego
}
\keyword{misc}

