% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make.response.mat}
\alias{make.response.mat}
\title{Converts a vector into a matrix where each column is a moving window}
\usage{
make.response.mat(X, bw = 5)
}
\arguments{
\item{X}{The vector to be processed}

\item{bw}{The bandwidth of the moving window, where the length of the
moving window is \code{2*bw + 1}.}
}
\value{
A matrix, where each column contains the elements of the
progressing moving window across \code{X}.
}
\description{
Converts a vector into a matrix where each column is a moving window
}
\details{
Near the beginning and end of the series, the moving window is padded with
\code{NA}'s as necessary.
}
\examples{
make.response.mat(rnorm(10), bw = 4)
}
\author{
Landon Sego
}
\seealso{
This function is used in \code{\link{calcIVTslow}}
}
\keyword{misc}

