% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sp.table.R
\name{sp.table}
\alias{sp.table}
\title{A 'special' table function optimized for integers}
\usage{
sp.table(y, y.unique, pct = FALSE)
}
\arguments{
\item{y}{An integer vector}

\item{y.unique}{The unique values of \code{y} whose occurence in \code{y}
will be counted}

\item{pct}{\code{=TRUE} report the counts as percentages}
}
\value{
An integer vector containing the counts or percentages of each of
the elements of \code{y.unique}
}
\description{
A function similar to the \code{\link{table}} function that works only for
integers for added speed.
}
\details{
\code{y.unique} may contain values that are not in \code{y}.
}
\examples{
x <- sample(c(rep(1,5), rep(2,3)))
sp.table(x, 1:2)
sp.table(x, 1:3)
}
\author{
Landon Sego
}
\seealso{
\code{\link{table}}
}

