% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summaryStats}
\alias{summaryStats}
\title{Create a function that calculates the summary statistics}
\usage{
summaryStats(stats)
}
\arguments{
\item{stats}{A character vector of summary statistics.  See details for acceptable
values of \code{stats}}
}
\value{
A function that takes a single argument of a numerical vector, and returns
a named vector of the summary statistics of that vector
}
\description{
Create a customized function that calculates desired summary statistics for a vector
of numerical input
}
\details{
The acceptable values of \code{stats} are
\describe{
 \item{"min"}{minimum}
 \item{"q1"}{first quartile (25th percentile)}
 \item{"mean"}{mean}
 \item{"med"}{median (50th percentile)}
 \item{"q3"}{third quartile (75th percentile)}
 \item{"max"}{maximum}
 \item{"sd"}{standard deviation}
 \item{"sum"}{sum}
 \item{"ss"}{sum of squares}
 \item{"count"}{number of non-missing elements in the vector}
 \item{"skew}{skewness}
 \item{"kurt"}{kurtosis (Pearson's measure)}
}
All statistics are calcuated such that if there are too many \code{NAs}, an \code{NA}
is returned.
}
\examples{
f <- summaryStats(c("q3","q1","mean","count","med"))
f
f(rnorm(200))
}
\author{
Landon Sego
}

