% Generated by roxygen2 (4.0.1): do not edit by hand
\name{transMap}
\alias{transMap}
\title{Summarizes the type and location of transitions between discrete states in
an integer vector}
\usage{
transMap(y, y.unique)
}
\arguments{
\item{y}{The integer vector whose state transitions will be identified}

\item{y.unique}{An integer vector containing the unique values of \code{y}}
}
\value{
A list with the following components: (also see example below)
\item{diffMap}{The first differences of \code{m(y)} which identify the
position and type of state transition in \code{y}} \item{from}{The 'from'
state labels} \item{to}{The 'to' state labels} \item{transValue}{The value
that \code{diffMap} will take when \code{y} transitions from the state
label in \code{from} to the corresponding state label in \code{to}}
}
\description{
Summarizes the type and location of transitions between discrete states in
an integer vector
}
\details{
The algorithm applies a discrete, 1-1 map, say \code{m}, to the vector
\code{y} such that the first differences of m(y) uniquely identify the
transitions between all the possible values of \code{y}. It creates the
map, calculates the first differences of \code{m(y)}, and produces a set of
transition labels that indicate which value of \code{m(Y)} corresponds to
each of the possible transitions among the unique states of \code{y}.
}
\examples{
transMap(c(1,1,1,2,2,1,1,3,3,2,2), 1:3)

# These are the mapped values which signal when
# a transition took place from one state to another
# and they uniquely identify which type of transistion
# it was.  Their values can be decoded using 'transValue'
# below.  A value of 0 indicates that no transition occured
# at that index.
# $diffMap
#  [1]  0  0  0 -1  0  1  0 -3  0  2  0

# The 'from' state labels
# $from
# [1] 1 1 2 2 3 3

# The 'to' state labels
# $to
# [1] 2 3 1 3 1 2

# The transistion value key.  e.g., -1 corresponds to a
# transition from state 1 to state 2.  -3 corresponds
# to a transition from state 1 to state 3.  etc.
# $transValue
# [1] -1 -3  1 -2  3  2
}
\author{
Landon Sego
}
\seealso{
This function is called in \code{\link{discreteTransition}}
}
\keyword{misc}

