% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sp.min}
\alias{sp.count}
\alias{sp.max}
\alias{sp.min}
\alias{sp.sum}
\title{Adaptations of common summary functions to return NA if all elements are NA}
\usage{
sp.min(x)
sp.max(x)
sp.sum(x)
sp.count(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
\code{sp.min}, \code{sp.max}, \code{sp.sum} operate just like
\code{\link{min}}, \code{\link{max}}, and \code{\link{sum}} except they
return \code{NA} if all elements of \code{x} are \code{NA}.

\code{sp.count} returns the number of non-missing values in the vector.
However, if all values are missing, it returns \code{NA}.
}
\description{
Adaptations of \code{\link{min}}, \code{\link{max}}, and \code{\link{sum}}
to return \code{NA} if all elements are \code{NA}.  The prefix 'sp' could
be thought of as 'special'.
}
\examples{
min(NA, na.rm=TRUE)
sp.min(NA)

max(NA, na.rm=TRUE)
sp.max(NA)

sum(NA, na.rm=TRUE)
sp.sum(NA)

sp.count(c(1, 3, NA, 4))
sp.count(rep(NA, 3))
}
\author{
Landon Sego
}
\keyword{misc}

