% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fitQ.R
\name{fitQ}
\alias{fitQ}
\alias{summary.fitQ}
\title{Fits a moving window quadratic (or simple linear) regression model}
\usage{
fitQ(y, x1 = -10:10, min.window = 5, start = 1, skip = 1,
  linear.only = FALSE)

\method{summary}{fitQ}(fitQ_object, stats = c("min", "q1", "mean", "med",
  "q3", "max", "sd", "count"))
}
\arguments{
\item{y}{A numeric vector}

\item{x1}{The predictor of the regression model, whose length must
be odd, and it must be monotonic (increasing or decreasing).  Typically this
would be an evenly spaced, increasing vector.}

\item{min.window}{The minimum number of non-missing data points in a window
that are required to fit the regression model.}

\item{start}{The index of the center of the first window}

\item{skip}{The number of indexes to advance the center of the moving
window each time the model is fit.}

\item{linear.only}{\code{=TRUE} fits a simple linear regression model with
\code{x1} as the single predictor, instead of a quadratic regression model.}

\item{stats}{A character vector of summary statistics that are valid for
\code{\link{summaryStats}}. Alternatively, the object returned by
\code{\link{summaryStats}} may also be supplied for this argument.}
}
\value{
An object of class \code{fitQ}, which is list with
the 4 vectors that contain the results of the
quadratic model fits:
\item{a}{The estimated intercepts}
\item{b}{The estimated linear coefficients}
\item{c}{The estimated quadratic coefficients. These are \code{NA} is
         \code{linear.only = TRUE}}
\item{d}{The root mean squared error (RMSE)}
The first element of the vectors \code{a}, \code{b}, \code{c}, and \code{d} contains the model
coefficients corresponding to the first data point in \code{y}.  The second element in
the vectors \code{a}, \code{b}, \code{c}, and \code{d} contains the model
coefficients corresponding to the second data point in \code{y}, etc.
}
\description{
Fits a moving quadratic (or simple linear) regression model over a series
using a 2-sided window.
}
\details{
This function dynamically accounts for the incomplete windows
which are caused by missing values and which occur at the beginning and end
of the series.
A quadratic regression model is used if \code{linear.only = FALSE}:

\code{y = a + b*x1 + c*x1^2 + e}

and a simple linear regression model is used if \code{linear.only = TRUE}:

\code{y = a + b*x1 + e}

where \code{a}, \code{b}, and \code{c} are regression coefficients, and \code{e}
represents the residual error.  The regression model is fit repeatedly over a two-sided
moving window across the values of \code{y}. The middle element of \code{x1} is aligned with
an element of \code{y} and then the regression model coefficients are calculated for that element
of \code{y}.
Consequently, the size of the moving window is determined by
the the number of elements in \code{x1}. For example, if \code{x1 = -3:3}, the size
of the moving window is 7.

For the quadratic model, \code{fitQ} rapidly calculates the least squares estimates by centering the
predictors so that they are orthogonal (or nearly orthogonal, if data in
the window are missing).  Then the orthogonal coefficients are
'backtransformed' to the original, non-orthogonal parameterization.
Centering and orthogonality are not required to quickly fit the simple
linear regression model.

The handling of the moving windows and missing values in this function is
very similar to \code{\link[Smisc:smartFilter]{Smisc::smartFilter}}.

Instead of a numeric vector, the \code{x1} argument can be a
\code{valid_fitQ_args} object (returned by \code{\link{check_fitQ_args}}), in which
case all the subsequent arguments to \code{fitQ} are ignored
(because the \code{valid_fitQ_args} object contains all those arguments). This is useful
because \code{fitQ} is called repeatedly by \code{\link{getFeatues}} over the same set of
argument values.
}
\section{Methods (by generic)}{
\itemize{
\item \code{summary}: Calculates summary statistics for a \code{fitQ} object, returning a
named vector with summary statistics.  The names take the form [coefficient].[stat],
where the coefficient is one of "a", "b", "c", or "d", and the stat is the statistic
required in the \code{stats} argument of the \code{summary} method.
}}
\examples{
# Calculate the rolling window quadratic fit
z <- fitQ(rnorm(25), -3:3)
z
summary(z)

# Or we can request our own summary stats
summary(z, stats = c("count", "mean", "kurt"))
}
\author{
Landon Sego
}

