% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getFeatures.R
\name{getFeatures}
\alias{getFeatures}
\title{Compute the quadratic, linear, and/or discrete features of multiple variables for a single group}
\usage{
getFeatures(y, cont = NULL, disc = NULL, centerScale = TRUE,
  stats = c("min", "q1", "mean", "med", "q3", "max", "sd", "count"),
  fitQargs = NULL)
}
\arguments{
\item{y}{Data frame, each row containing a vector of measurements for a particular point in time,
with columns indicating the measured variables (and possibly other descriptive variables).  The
data processed presuming the rows are orderd chronologically.}

\item{cont}{Vector of integers or a character vector indicating the columns
of \code{x} that correspond to continuous variables.  These are the variables from which features
will be extracted by fitting the moving regression model using \code{\link{fitQ}}.}

\item{disc}{Vector of integers or character vector indicating the columns
of \code{x} that correspond to variables that will be treated as discrete. These are the variables
from which features will be extracted using \code{\link{discFeatures}}.}

\item{centerScale}{Logical indicating whether the continuous variables (indicate by \code{cont})
should be centered and scaled by the global mean and standard deviation of that variable.  By
'global', we mean all the values of a continuous variable, say \code{x}, in \code{y} are used
to compute the mean and standard deviation.  The resulting value for the continuous variable,
\code{x}, is equivalent to \code{y$x <- (y$x - mean(y$x)) / sd(y$x)}.}

\item{stats}{This argument defines the summary statistics that will be calculated
for each of the regression parameters.  It can be a character vector of summary statistics,
which are passed to \code{\link{summaryStats}}.  Or the function object returned by
\code{\link{summaryStats}} may be supplied.}

\item{fitQargs}{Named list of arguments for \code{\link{fitQ}}. If \code{NULL}, the default
arguments of \code{\link{fitQ}} are used. Any argument for \code{\link{fitQ}} may be included
except \code{y}.}
}
\value{
A named vector containing the features for each of the variables
requested in \code{cont} and \code{disc}.  The names follow the form
[varname].[description], where the [varname] is specified in \code{cont} and
\code{disc}, and [description] follows the naming convention produced by
\code{\link{summary.fitQ}} and \code{\link{discFeatures}}.
}
\description{
Fits the moving window quadratic (or linear) regression model for each continuous variable in
a data frame, and calculates summary statistics of the parameters.  Also calculates the
duration and transition features of discrete variables.
}
\details{
A least one of \code{cont} or \code{disc} must be specified.

Instead of a data frame, the \code{y} argument can be a
\code{valid_getFeatures_args} object (returned by \code{\link{check_getFeatures_args}}), in which
case all the subsequent arguments to \code{getFeature} are ignored
(because the \code{valid_getFeatures_args} object contains all those arguments). This is useful
if \code{getFeatures} is called repeatedly over the same set of argument values (which occurs
in \code{\link{ddply_getFeatures}}.
}
\examples{
# Load the data
data(demoData)

# Select a subset of thedata
d <- demoData[demoData$subject == 3 & demoData$phase == "f",]
colnames(d)

# Run over that subset
features <- getFeatures(d, cont = 3:4, disc = 8:11, stats = c("mean", "sd"),
                        fitQargs = list(x1 = -5:5, start = 2))
str(features)
features

# We can also call the function by validating the arguments before hand:
validated <- check_getFeatures_args(d, cont = 3:4, disc = 8:11, stats = c("mean", "sd"),
                                    fitQargs = list(x1 = -5:5, start = 2))

features1 <- getFeatures(validated)

# We get the same result
identical(features1, features)
}
\author{
Landon Sego
}
\references{
Amidan BG, Ferryman TA. 2005.  “Atypical Event and Typical Pattern Detection within Complex Systems.”
IEEE Aerospace Conference Proceedings, March 2005.
}

