% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qFeature-package.R
\docType{package}
\name{qFeature-package}
\alias{qFeature}
\alias{qFeature-package}
\title{Extract features from time series using moving windows of regression fits}
\description{
Extract features from time series using moving windows of regression fits.This is a
high-speed implementation of the feature extraction methods of the
Morning Report Algorithms developed by Brett Amidan and Tom Ferryman.
}
\details{
\tabular{ll}{
Package: \tab qFeature\cr
Type: \tab Package\cr
Version: \tab 0.1.2\cr
Date: \tab 2015-02-16\cr
License: \tab LICENSE\cr }
}
\examples{
help(package = qFeature)
}
\author{
Landon Sego

Maintainer: Landon Sego <Landon.Sego@pnnl.gov>
}
\keyword{package}

