% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qFeature-package.R
\docType{package}
\name{qFeature-package}
\alias{qFeature}
\alias{qFeature-package}
\title{Extract Features from Continuous or Discrete Time Series}
\description{
Extract features from time series using moving windows of regression fits.
This is a high-speed implementation of the feature extraction methods of the
Morning Report Algorithms developed by Brett Amidan and Tom Ferryman.
}
\details{
\tabular{ll}{
Package: \tab qFeature\cr
Type: \tab Package\cr
Version: \tab 0.1.4\cr
Date: \tab 2015-09-09\cr
License: \tab file LICENSE\cr }
}
\examples{
\dontrun{
help(package = qFeature)
browseVignettes("qFeature")
}
}
\author{
Landon Sego, Lucas Tate

Maintainer: Landon Sego <Landon.Sego@pnnl.gov>
}
\references{
Amidan BG, Ferryman TA. 2005.  "Atypical Event and Typical Pattern Detection within Complex Systems."
IEEE Aerospace Conference Proceedings, March 2005.
}
\keyword{package}

