% Generated by roxygen2 (4.0.2): do not edit by hand
\name{qFeature}
\alias{qFeature}
\title{Compute the quadratic, linear, and/or discrete features of multiple variables for a single group}
\usage{
qFeature(y, .variables, cont = NULL, disc = NULL, stats = c("min", "q1",
  "mean", "med", "q3", "max", "sd", "count"), fitQargs = NULL, ...)
}
\arguments{
\item{.variables}{variables (i.e. columns) in \code{y} to split data frame by, input as 'as.quoted' variables.
These combinations of the variables uniquely identify the groups for which the features will be
separately extracted.  This is passed directly to the argument of the same name in
\code{\link[plyr:ddply]{plyr::ddply}}.}

\item{cont}{Vector of integers or a character vector indicating the columns
of \code{x} that correspond to continuous variables.  These are the variables from which features will
be extracted by fitting the moving regression model using \code{\link{fitQ}}.}

\item{disc}{Vector of integers or character vector indicating the columns
of \code{x} that correspond to variables that will be treated as discrete. These are the variables
from which features will be extracted using \code{\link{discFeatures}}.}

\item{stats}{This argument defines the summary statistics that will be calculated
for each of the regression parameters.  It can be a character vector of summary statistics,
which are passed to \code{\link{summaryStats}}.  Or the function object returned by
\code{\link{summaryStats}} may be supplied.}

\item{fitQargs}{Named list of arguments for \code{\link{fitQ}}.  If \code{NULL}, the default arguments of
\code{\link{fitQ}} are used.}

\item{x}{Data frame, each row containing a vector of measurements for a particular point in time, with
columns indicating the measured variables (and possibly other descriptive variables).  The
data processed presuming the rows are orderd chronologically.}

\item{\dots}{Additional arguments to \code{\link[plyr:ddply]{plyr::ddply}}.}
}
\value{
A dataframe with one row for each grouping defined by \code{.variables}.  The features computed by
\code{\link{getFeatures}} is presented across the columns.
}
\description{
This is a wrapper that implements \code{\link{getFeatures}} for each group in a data frame using
\code{\link[plyr:ddply]{plyr::ddply}}.
}
\details{
A least one of \code{cont} or \code{disc} must be specified.
}
\author{
Landon Sego
}

