% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sp.min}
\alias{sp.count}
\alias{sp.kurtosis}
\alias{sp.max}
\alias{sp.mean}
\alias{sp.min}
\alias{sp.skewness}
\alias{sp.sum}
\title{Adaptations of common summary functions to return NA if all elements are NA}
\usage{
sp.min(x)
sp.max(x)
sp.sum(x)
sp.count(x)
sp.mean(x)
sp.skewness(x)
sp.kurtosis(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
\code{sp.min}, \code{sp.max}, \code{sp.sum} operate just like
\code{\link{min}}, \code{\link{max}}, and \code{\link{sum}} except they
return \code{NA} if all elements of \code{x} are \code{NA}.

\code{sp.count} returns the number of non-missing values in the vector.
However, if all values are missing, it returns \code{NA}.
}
\description{
Adaptations of common summary statistical functions
to return \code{NA} if all elements of \code{x} are \code{NA}.  The prefix 'sp' could
be thought of as 'special'. These are internal functions used by
\code{\link{summaryStats}}.
}
\author{
Landon Sego
}
\keyword{misc}

